Add stop-the-world garbage collection
We use a mark-and-sweep GC that stops execution when collecting. The tri-color marking algorithm employs a DFS search that initially marks black nodes children (nodes that are marked not to be sweeped) to be gray, then look at the grey children to mark their children gray which, in turn, marks that node black. We then sweep. GC is done to maximise throughput (ratio of user code run and GC) and minimise latency (average time spent in GC). We collect whenever we encounter the memory allocated to be larger than a target GC memory size.